python_api/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Validation / Settings
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.1

# Payments
stripe==9.6.0

# HTTP client (webhooks / external calls)
httpx==0.27.0

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.2.0

# Dev / testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0
